agentcore-identity/pyproject.toml
[project]
name = "agentcore-identity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bedrock-agentcore>=1.0.5",
    "boto3>=1.40.65",
    "strands-agents>=1.14.0",
]
